commit to git (source)
hope this doesn't fucking you know what i eman

+++ src/ethos/model/npcs/NPCDeathTracker.java
@@ -45,6 +45,9 @@ public class NPCDeathTracker {
 			if (name.equalsIgnoreCase("none")) {
 				return;
 			}
+			if (name.equalsIgnoreCase("olm")) {
+				return;
+			}
 			if (!name.contains("clue-scroll")) {
 				player.sendMessage("Your " + Misc.capitalizeJustFirst(name.replaceAll("_", " "))
 					+ " kill count is: <col=FF0000>" + kills + "</col>.");
